requirements.txt
streamlit==1.45.1
langchain==0.3.25
langchain-openai==0.3.17
langchain-community==0.3.24
faiss-cpu==1.11.0
pypdf==5.4.0
python-dotenv==1.1.0
tiktoken==0.9.0
